requirements.txt
fastapi
uvicorn
pymongo
python-decouple
openai